Pretty much the same as the unset one, minus maximizing any parts (removing maximizations is a requirement, though).
I need to decide whether to put these into separate functions. The
code that's the same is to get multiple values and, without the
ability to pass said values back, that means the function would likely
need to do the `wmctrl` command, as well – but that's the part that's
different.

I could try passing boolean values to determine which version to run
but that ties them down to the process being run rather than any sort
of inherent or internal logic.

If I could narrow each final variable to a function, then I could just
call those directly from the `wmctrl` calls? It'd reduce things, for
sure, but some of them require the same, other variable, which results
in the function often only being a single command. Sure, it's good
practice to have any repeated code in a function so changes are only
made once but, for one line, not strictly needed unless the lines is
called in, at least, three places (plus I'm not certain it's worth the
loss of immediately logic-following for new readers that comes from
the logic being in the place of use rather than having to remove
oneself to find the function being called and then resume following
the logic…).