We're gonna switch over to LXTerminal since URxvt seems to behave weirdly with unicode and I can't seem to get it to just display properly.
LXTerminal, on the other hand, just works; plus it seems to take less
RAM with multiple terminals running, even. On top of all that, it has
an intuitive UI that most users will be able to recognize.

I'll miss the level of control URxvt offered (namely being able to
control things like decoration or scrollbars from a terminal command)
but I can live with that for things just working and not getting weird
unicode boxes unexpectedly.