Update all locations.
Another plus is we can generalize this part so, if the user ever
changes terminals, we won't break their config.

I'm not sure if it's because LXTerminal takes longer to load but it
keeps not always coming to the forefront for any command that launches
it; I hate applying such heavy-handed settings to every app. and
messing with app.s native behavior but not having new windows go to
the forefront  is going to get hella old real fast…